file reformat and functions placement
